+++ src/components/PageNotFound.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const PageNotFound = () => {
+	return (
+		<div>
+			404 Page Not Found!
+		</div>
+	);
+};
+
+export default PageNotFound;
