+++ src/components/Profile.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Profile = () => {
+	return (
+		<div>
+			This is Page Profile
+		</div>
+	);
+};
+
+export default Profile;
