--- src/components/chats.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const Chats = () => {
-    return (
-        <div>
-            
-        </div>
-    );
-};
-
-export default Chats;
